refactor: Long to Integer user count

+++ spring-boot/src/main/java/org/springframework/ntfh/entity/user/UserController.java
@@ -68,8 +68,8 @@ public class UserController {
 	}
 
 	@GetMapping("count")
-	public ResponseEntity<Long> getCount() {
-		Long userCount = this.userService.count();
+	public ResponseEntity<Integer> getCount() {
+		Integer userCount = this.userService.count();
 		return new ResponseEntity<>(userCount, HttpStatus.OK);
 	}
